Fixed some warnings.

+++ radrenderer/src/core/Window.cpp
@@ -133,7 +133,7 @@ void Window::loop()
 
         sf::Time elapsed = m_clock.restart();
 
-        std::uint8_t* pixels = reinterpret_cast<std::uint8_t*>(m_renderer.update(elapsed.asMilliseconds(), forward, rotate_x, rotate_y));
+        std::uint8_t* pixels = reinterpret_cast<std::uint8_t*>(m_renderer.update((float)elapsed.asMilliseconds(), forward, rotate_x, rotate_y));
 
         float fps = 1.f / elapsed.asSeconds();
